Implemented time saves for pathfinding.
