+++ ewmap.py
@@ -19,26 +19,6 @@ from ewplayer import EwPlayer
 
 move_counter = 0
 
-def get_move_speed(user_data):
-	mutations = user_data.get_mutations()
-	market_data = EwMarket(id_server = user_data.id_server)
-	move_speed = 1
-        
-	if user_data.life_state == ewcfg.life_state_corpse:
-		move_speed *= 0.5
-
-	if ewcfg.mutation_id_organicfursuit in mutations and (
-		(market_data.day % 31 == 0 and market_data.clock >= 20)
-		or (market_data.day % 31 == 1 and market_data.clock < 6)
-	):
-		move_speed *= 2
-	if ewcfg.mutation_id_lightasafeather in mutations and market_data.weather == "windy":
-		move_speed *= 2
-	if ewcfg.mutation_id_fastmetabolism in mutations and user_data.hunger / user_data.get_hunger_max() < 0.5:
-		move_speed *= 2
-
-	return move_speed
-
 
 """
 	Returns true if the specified point of interest is a PvP zone.
@@ -379,7 +359,7 @@ def path_step(path, coord_next, user_data, coord_end):
 
 	path.steps.append(coord_next)
 
-	cost_next = int(cost_next / get_move_speed(user_data))
+	cost_next = int(cost_next / user_data.move_speed)
 
 	path.cost += cost_next
 
@@ -780,14 +760,14 @@ async def move(cmd):
 								boost = ewcfg.territory_time_gain
 							else:
 								territory_slowdown = ewcfg.territory_time_gain
-								territory_slowdown = int(territory_slowdown / get_move_speed(user_data))
+								territory_slowdown = int(territory_slowdown / user_data.move_speed)
 								await asyncio.sleep(territory_slowdown)
 			else:
 				if val > 0:
 					val_actual = val - boost
 					boost = 0
 
-					val_actual = int(val_actual / get_move_speed(user_data))
+					val_actual = int(val_actual / user_data.move_speed)
 					if val_actual > 0:
 						await asyncio.sleep(val_actual)
